agentmesh-eda/pyproject.toml
[tool.poetry]
name = "agentmesh-eda"
version = "0.1.0"
description = "Event-Driven Architecture for Agentic AI"
authors = ["Gemini"]
packages = [{include = "agentmesh"}]

[tool.poetry.dependencies]
python = "^3.9"
confluent-kafka = "^2.3.0"
nats-py = "^2.3.1"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
google-cloud-pubsub = "^2.19.0"
boto3 = "^1.34.5"
pulsar-client = "^3.4.0"
loguru = "^0.7.2"
prometheus-client = "^0.20.0"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.5"
sphinx = "^7.2.6"
ruff = "^0.1.11"

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
